Add matlab 2021a
I would go for a newer release, but due to some licensing stuff
that doesn't concern me, this is the latest release where i can get
the license file

To obtain the license file, sign in to mathworks, go to license center,
select your license, under install tab, there should be clickable
'activate a computer' thing, and from there you can download the license.

Would love to automate that but i don't use that many machines (yet) for it
to be a problem.

Now adding a license requires you to know a host id. The way i got that id
is i just copied a random license file into the container,
and the error when running matlab let me know what the proper id was.

According to the internet, the mathworks host id is just the mac address,
so that might be worth exploring if i were to ever automate this process.